restructure causal analysis
