+++ src/dackar/causal/OperatorShiftLogsProcessing.py
@@ -15,7 +15,7 @@ from spacy.tokens import Span
 from ..text_processing.Preprocessing import Preprocessing
 from ..utils.utils import getOnlyWords, getShortAcronym
 from ..config import nlpConfig
-from .WorkflowBase import WorkflowBase
+from .CausalBase import CausalBase
 from ..pipelines.CustomPipelineComponents import mergeEntitiesWithSameID
 
 logger = logging.getLogger(__name__)
@@ -50,7 +50,7 @@ if not Token.has_extension('action'):
 if not Token.has_extension('edep'):
   Token.set_extension("edep", default=None)
 
-class OperatorShiftLogs(WorkflowBase):
+class OperatorShiftLogs(CausalBase):
   """
     Class to process OPG Operator Shift Logs dataset
   """
@@ -351,74 +351,6 @@ class OperatorShiftLogs(WorkflowBase):
           # if the entity not among subj and obj and there are more than one entity, it may not need to report it
           else:
             pass
-
-  def extractRelDep(self, matchedSents):
-    """
-
-      Args:
-
-        matchedSents: list, the list of matched sentences
-
-      Returns:
-
-        (subject tuple, predicate, object tuple): generator, the extracted causal relation
-    """
-    subjList = ['nsubj', 'nsubjpass', 'nsubj:pass']
-    objList = ['pobj', 'dobj', 'iobj', 'obj', 'obl', 'oprd']
-    for sent in matchedSents:
-      ents = self.getCustomEnts(sent.ents, self._entityLabels[self._entID])
-      if ents is None or len(ents) <= 1:
-        continue
-      root = sent.root
-      allRelPairs = []
-      subjEnt = []
-      subjConjEnt = []
-      objEnt = []
-      objConjEnt = []
-
-      for ent in ents:
-        entRoot = ent.root
-        if ent._.alias is not None:
-          text = ent._.alias
-        else:
-          text = ent.text
-        # entity at the beginning of sentence
-        if ent.start == sent.start:
-          subjEnt.append(text)
-        elif entRoot.dep_ in ['conj'] and entRoot.i < root.i:
-          subjConjEnt.append(text)
-        elif entRoot.dep_ in subjList:
-          subjEnt.append(text)
-        # elif entRoot.dep_ in ['obj', 'dobj']: # mainly for short sentence or phrase
-        elif entRoot.dep_ in objList:
-          objEnt.append(text)
-        elif entRoot.i > root.i and entRoot.dep_ in ['conj']:
-          objConjEnt.append(text)
-      # subj
-      for subj in subjEnt:
-        for subjConj in subjConjEnt:
-          allRelPairs.append([subj, 'conj', subjConj])
-        for obj in objEnt:
-          allRelPairs.append([subj, root, obj])
-        for objConj in objConjEnt:
-          allRelPairs.append([subj, root, objConj])
-      # subjconj
-      for subjConj in subjConjEnt:
-        for obj in objEnt:
-          allRelPairs.append([subjConj, root, obj])
-        for objConj in objConjEnt:
-          allRelPairs.append([subjConj, root, objConj])
-      # obj
-      for obj in objEnt:
-        for objConj in objConjEnt:
-          allRelPairs.append([obj, 'conj', objConj])
-
-      # handle specific case
-      if len(allRelPairs) == 0:
-        if len(ents) == 2 and ents[0].root.i < root.i and ents[1].root.i > root.i:
-          allRelPairs.append([ents[0].text, root.text, ents[1].text])
-
-      self._allRelPairs += allRelPairs
 
   def extractCausalRelDep(self, matchedSents):
     """
